增加修改失败的处理
