+++ apps/mis-web/src/pages/api/profile/changeEmail.ts
@@ -11,10 +11,16 @@
  */
 
 import { typeboxRoute, typeboxRouteSchema } from "@ddadaal/next-typed-api-routes-runtime";
-import { changeEmail as libChangeEmail } from "@scow/lib-auth";
+import { asyncClientCall } from "@ddadaal/tsgrpc-client";
+import { Status } from "@grpc/grpc-js/build/src/constants";
+import { UserServiceClient } from "@scow/protos/build/server/user";
 import { Type } from "@sinclair/typebox";
 import { authenticate } from "src/auth/server";
-import { publicConfig, runtimeConfig } from "src/utils/config";
+import { getClient } from "src/utils/client";
+import { publicConfig } from "src/utils/config";
+import { handlegRPCError } from "src/utils/server";
+
+
 
 // 此API用于用户修改自己的邮箱。
 export const ChangeEmailSchema = typeboxRouteSchema({
@@ -22,6 +28,7 @@ export const ChangeEmailSchema = typeboxRouteSchema({
   method: "PATCH",
 
   body: Type.Object({
+    userId:Type.String(),
     newEmail: Type.String(),
   }),
 
@@ -31,6 +38,9 @@ export const ChangeEmailSchema = typeboxRouteSchema({
 
     /** 用户未找到 */
     404: Type.Null(),
+
+    /** 修改失败 */
+    500: Type.Null(),
 
     /** 本功能在当前配置下不可用。 */
     501: Type.Null(),
@@ -49,26 +59,17 @@ export default /* #__PURE__*/typeboxRoute(ChangeEmailSchema, async (req, res) =>
 
   if (!info) { return; }
 
-  const { newEmail } = req.body;
+  const { userId, newEmail } = req.body;
 
+  const client = getClient(UserServiceClient);
 
-  return await libChangeEmail(runtimeConfig.AUTH_INTERNAL_URL, {
-    identityId: info.identityId,
+  return await asyncClientCall(client, "changeEmail", {
+    userId,
     newEmail,
-  }, console)
-    .then(() => {
-      return ({ 204: null });
-    })
-    .catch((e) => {
-      switch (e.status) {
-      case "NOT_FOUND":
-        return { 404: null };
-      case "NOT_SUPPORTED":
-        return { 501: null };
-      default:
-        throw e;
-      }
-    });
-
-
+  })
+    .then(() => ({ 204: null }))
+    .catch(handlegRPCError({
+      [Status.NOT_FOUND]: () => ({ 404: null }),
+      [Status.UNKNOWN]: () => ({ 500: null }),
+    }));
 });
